requirements.txt
alembic==1.4.2
astroid==2.4.1
click==7.1.2
Flask==1.1.2
Flask-API==2.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.17
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
